refactor : 채팅 면접 진행 상태 provider 로직 일부 수정
