+++ lib/features/chat/entities/enums/chat_room_progress.enum.dart
@@ -1,0 +1,14 @@
+///
+/// 면접(채팅방) 진행상태
+///
+enum ChatRoomProgress {
+  initial,
+  ongoing,
+  completed;
+
+  bool get isInitial => this == ChatRoomProgress.initial;
+  bool get isOngoing => this == ChatRoomProgress.ongoing;
+  bool get isCompleted => this == ChatRoomProgress.completed;
+  bool get isOnGoingOrCompleted =>
+      this == ChatRoomProgress.ongoing || this == ChatRoomProgress.completed;
+}
